Add files via upload
This workflow will automatically collect top 5 technology news and automatically upload it to your google sheets for future purposes. After that it will send an summary of the news through your Gmail.  This workflow will keeps you uptodate at Technology, Political ,Economy. If you want to change the no of news you can do that on the if node